refactor(eigensolver): putting more work into krylov base class
the idea is to make use of its tolerance and stop computation
if the required evs have converged
