+++ Solvers/EigenSolvers/CMakeLists.txt
@@ -7,6 +7,7 @@ target_include_directories(pz PUBLIC
 
 set(public_headers
     TPZLapackEigenSolver.h
+    TPZKrylovEigenSolverBase.h
     TPZKrylovEigenSolver.h
     TPZLinearEigenSolver.h
     TPZQuadEigenSolver.h
@@ -18,6 +19,7 @@ set(headers
 
 set(sources
     TPZLapackEigenSolver.cpp
+    TPZKrylovEigenSolverBase.cpp
     TPZKrylovEigenSolver.cpp
     TPZLinearEigenSolver.cpp
     TPZQuadEigenSolver.cpp
