+++ Solvers/EigenSolvers/TPZKrylovEigenSolver.cpp
@@ -23,10 +23,10 @@ void TPZKrylovEigenSolver<TVar>::AdjustTargetST()
   }
 }
 
+
+
 template<class TVar>
-int TPZKrylovEigenSolver<TVar>::SolveImpl(TPZVec<CTVar> &w,
-                                          TPZFMatrix<CTVar> &eigenVectors,
-                                          bool computeVectors)
+void TPZKrylovEigenSolver<TVar>::PreSolve()
 {
 #ifdef USING_MKL
   auto PardisoSetup = [](auto pardiso_control, auto sys, auto prop){
@@ -82,7 +82,7 @@ int TPZKrylovEigenSolver<TVar>::SolveImpl(TPZVec<CTVar> &w,
   PZError<<"\nERROR: NeoPZ was not linked against LAPACK. Aborting...\n";
   DebugStop();
 #endif
-  TPZSimpleTimer total("Arnoldi Solver",true);
+
 
   
   const int64_t nRows = this->MatrixA()->Rows();
@@ -131,87 +131,41 @@ int TPZKrylovEigenSolver<TVar>::SolveImpl(TPZVec<CTVar> &w,
       else this->MatrixB()->Decompose(ELDLt);
     }
   }
-  
-  const int &n = this->NEigenpairs();
-  if(this->KrylovDim() == -1){
-    this->SetKrylovDim(10*n);
-  }
-  const int &krylovDim = this->KrylovDim();
-  TPZManVector<TPZAutoPointer<TPZFMatrix<TVar>>,20> qVecs;
-  TPZFNMatrix<400,TVar> h(krylovDim,krylovDim,0.);
 
-  
-  auto success = this->ArnoldiIteration(qVecs,h);
-  if(!success){
-    return -1;
-  }
-  TPZFNMatrix<400,CTVar> lapackEV(n,n,0.);
-  
+}
 
-  auto lapackres = [&h,&w,&lapackEV]()
-  {
-    TPZSimpleTimer lapacktimer("Hessenberg EVP");
-    TPZLapackEigenSolver<TVar> lapack;
-    return lapack.SolveHessenbergEigenProblem(h, w, lapackEV);
-  }();
-  if(lapackres) return lapackres;
-
+template<class TVar>
+void TPZKrylovEigenSolver<TVar>::TransformEigenvalues(TPZVec<CTVar> &w)
+{
+  auto st = SpectralTransform();
   if(st) st->TransformEigenvalues(w);
-
-  TPZManVector<int,20> indices;
-  
-  this->SortEigenvalues(w,indices);
-  
-  if(!computeVectors) return lapackres;
-  eigenVectors.Redim(nRows,n);
-  {
-    TPZSimpleTimer evTimer("Computing eigenvectors");
-    for (int i = 0; i< n; i++){//which eigenvector from A
-      auto il = indices[i];
-      for (int j = 0; j < krylovDim; j++){//which vector from Q
-        CTVar *ev = &eigenVectors.g(0,i);
-        const auto lev = lapackEV(j,il);
-        TVar *q = &qVecs[j]->g(0,0);
-        /*
-          The following loop computes
-          eigenVectors(k,i) += lev * qvec.GetVal(k,0),
-          with const auto qvec = *qVecs[j],
-          but twice as fast
-        */
-        for(int k = 0; k < nRows; k++)
-          *ev++ += lev * *q++;
-      }
-    }
-  }
-  
-  return lapackres;
 }
 
 template<class TVar>
 int TPZKrylovEigenSolver<TVar>::SolveEigenProblem(TPZVec<CTVar> &w,TPZFMatrix<CTVar> &eigenVectors)
 {
-  return SolveImpl(w,eigenVectors,true);
+  return this->SolveImpl(w,eigenVectors,true);
 }
 
 template<class TVar>
 int TPZKrylovEigenSolver<TVar>::SolveEigenProblem(TPZVec<CTVar> &w)
 {
   TPZFMatrix<CTVar> eigenVectors;
-  return SolveImpl(w,eigenVectors,false);
+  return this->SolveImpl(w,eigenVectors,false);
 }
 
 template<class TVar>
 int TPZKrylovEigenSolver<TVar>::SolveGeneralisedEigenProblem(TPZVec<CTVar> &w,
                                  TPZFMatrix<CTVar> &eigenVectors)
 {
-  return SolveImpl(w,eigenVectors,true);
+  return this->SolveImpl(w,eigenVectors,true);
 }
 
 template<class TVar>
 int TPZKrylovEigenSolver<TVar>::SolveGeneralisedEigenProblem(TPZVec<CTVar> &w)
 {
   TPZFMatrix<CTVar> eigenVectors;
-  return SolveImpl(w, eigenVectors,false);
+  return this->SolveImpl(w, eigenVectors,false);
 }
 
 
@@ -242,84 +196,9 @@ void TPZKrylovEigenSolver<TVar>::ApplyOperator(const TPZFMatrix<TVar> &x, TPZFMa
 }
 //! System size (number of rows)
 template<class TVar>
-int64_t TPZKrylovEigenSolver<TVar>::SystemSize() const
+int64_t TPZKrylovEigenSolver<TVar>::NRows() const
 {
   return this->fMatrixA->Rows();
-}
-
-
-template<class TVar>
-bool TPZKrylovEigenSolverBase<TVar>::ArnoldiIteration(
-  TPZVec<TPZAutoPointer<TPZFMatrix<TVar>>> &Q,
-  TPZFMatrix<TVar> &H)
-{
-
-  if(KrylovDim() < 2){
-    fKrylovDim = 10;
-  }
-  const int64_t nRows = this->SystemSize();
-  const int n = std::min(fKrylovDim,nRows);
-  std::cout<<"Calculating Krylov subspace of dimension "<<n<<'\n';
-  H.Redim(n,n);
-  Q.Resize(n, nullptr);
-
-  if(fKrylovVector.Rows() != nRows || fKrylovVector.Cols() != 1){
-    fKrylovVector.AutoFill(nRows,1,SymProp::NonSym);
-  }
-  
-  for(int i = 0; i < n; i++) Q[i]= new TPZFMatrix<TVar>;
-  
-  /*see Chapter 2 of slepc manual(EPS) or search for Arnoldi Iteration*/
-
-  //initializing first vector
-  *(Q[0]) = fKrylovVector * (TVar)(1./Norm(fKrylovVector));
-  
-  TPZSimpleTimer arnoldiIteration("ArnoldiIteration");
-  const auto &tol = Tolerance();
-
-  TPZFMatrix<TVar> w(nRows,1,0.);
-
-  for(auto k = 1; k < n+1; k++){
-    // TPZSimpleTimer arnoldiStep("step"+std::to_string(k),true);
-
-    //let us generate a first guess for w: w = A.q_{k-1}
-    this->ApplyOperator(*Q[k-1],w);
-
-    RTVar normW{1};
-    bool success = false;
-    /** after orthogonalising w.r.t. previous vectors (gram-schmidt)
-        we will then have w_k = Av_k - sum_j^k (h_{jk} v_j)
-    */
-    while(!success){
-      for(auto j = k-1; j >= 0; j--){
-        const auto& qj = *(Q[j]);
-        const auto dotqj = Dot(w,qj);
-        H.PutVal(j,k-1,dotqj);
-        w -= qj * dotqj;
-      }
-
-      normW = Norm(w);
-      if(normW > tol || k == n){
-        success = true;
-      }
-      else{
-        //generate random unit vector and try again
-        w.AutoFill(nRows,1,SymProp::NonSym);
-        w *= 1/Norm(w);
-      }
-    }
-
-    
-    if(k < n){
-      H.PutVal(k,k-1,normW);
-      w *= (TVar)1./normW;
-      (*(Q[k])) = w;
-    }
-
-    
-  }//for k
-
-  return true;
 }
 
 template class TPZKrylovEigenSolver<float>;
